# Conflicts:
#	android/TodokTodok/app/src/main/java/com/team/todoktodok/data/datasource/book/DefaultBookRemoteDataSource.kt
#	android/TodokTodok/app/src/main/java/com/team/todoktodok/data/datasource/discussion/DefaultDiscussionRemoteDataSource.kt
#	android/TodokTodok/app/src/main/java/com/team/todoktodok/data/datasource/notification/DefaultNotificationLocalDataSource.kt
#	android/TodokTodok/app/src/main/java/com/team/todoktodok/data/repository/DefaultBookRepository.kt
#	android/TodokTodok/app/src/main/java/com/team/todoktodok/data/repository/DefaultDiscussionRepository.kt
#	android/TodokTodok/app/src/main/java/com/team/todoktodok/data/repository/DefaultNotificationRepository.kt

+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/data/network/response/latest/LatestDiscussionResponse.kt
@@ -10,14 +10,14 @@ import kotlinx.serialization.Serializable
 data class LatestDiscussionResponse(
     val member: MemberResponse,
     val book: BookResponse,
-    val commentCount: Int,
+    val commentCount: Long,
     val discussionOpinion: String,
     val discussionTitle: String,
     val createdAt: String,
     val discussionId: Long,
     val isLikedByMe: Boolean,
-    val viewCount: Int,
-    val likeCount: Int,
+    val viewCount: Long,
+    val likeCount: Long,
 ) {
     fun toDomain(): Discussion =
         Discussion(
